+++ modules/core-config/local.tf
@@ -8,6 +8,11 @@ locals {
 
   az_count = length(var.availability_zones)
 
+  system_namespaces = [
+    "default",
+    "kube-system"
+  ]
+
   namespaces = [
     "cert-manager",
     "dns",
@@ -15,6 +20,13 @@ locals {
     "ingress-core-internal",
     "monitoring"
   ]
+
+  namespace_pod_security_labels = {
+    "pod-security.kubernetes.io/audit-version" = "v${var.cluster_version}"
+    "pod-security.kubernetes.io/warn-version"  = "v${var.cluster_version}"
+    "pod-security.kubernetes.io/audit"         = "baseline"
+    "pod-security.kubernetes.io/warn"          = "baseline"
+  }
 
   ingress_internal_core = merge(var.core_services_config.ingress_internal_core, var.core_services_config.ingress_internal_core.subdomain_suffix == null ? { subdomain_suffix = var.cluster_name } : {}, {
     annotations = {